+++ src/main/java/org/c3lang/intellij/docs/FuncDefinitionDocs.java
@@ -47,6 +47,7 @@ public final class FuncDefinitionDocs
 		builder.append(DocumentationMarkup.SECTIONS_START);
 		builder.append(DocumentationUtils.extractDescriptionTextFromDoc(docs)).append('\n');
 		DocumentationUtils.appendParamsSection(docs, builder, args);
+		DocumentationUtils.appendContractSections(docs, builder);
 		DocumentationUtils.appendReturnSection(docs, builder);
 		DocumentationUtils.appendFileSection(file, builder);
 		builder.append(DocumentationMarkup.SECTIONS_END);